已修复。mFrameCount = mFrames.size()（imageanimated.cpp:122）补上 static_cast<int>，与仓内既有约定（directorymanager.cpp 等统一用 static_cast<int>(container.size())）一致，消除 bugprone-narrowing-conversions 警告。帧数上限受内存限制不可能接近 INT_MAX，转换安全。 其它位置已核对：imageCount()/nextImageDelay() 返回 int，index < mFrames.size() 是比较不缩窄，均无同类警告。

+++ qimgv/sourcecontainers/imageanimated.cpp
@@ -119,7 +119,7 @@ std::shared_ptr<const QImage> ImageAnimated::decodeFrame(int index) {
 
     // 实际帧数少于 imageCount() 的扫描值时修正
     if (reachedEnd && mFrames.size() < mFrameCount)
-        mFrameCount = mFrames.size();
+        mFrameCount = static_cast<int>(mFrames.size());
 
     return (index < mFrames.size()) ? mFrames.at(index) : nullptr;
 }
